# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS API Documentation",
        "description": "API documentation for LMS Smart Notebook project",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/iframe-token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Generate JWT token for LMS iframe",
                "description": "Generate a JWT access token for a student to use in LMS iframe. Returns the token, expiration time (seconds), and issued timestamp.",
                "operationId": "df550ce038c80424f2a893de34101f3c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "course_id",
                                    "unit_number",
                                    "lesson_title"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "unit_number": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "lesson_title": {
                                        "type": "string",
                                        "example": "Introduction to JWT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "JWT token and session info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJK..."
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        },
                                        "issued_at": {
                                            "type": "integer",
                                            "example": 1716200000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/flashcards/generate": {
            "post": {
                "tags": [
                    "Flashcards"
                ],
                "summary": "Generate flashcards for a note using AI",
                "description": "Generates up to 5 educational flashcards for the specified note using OpenAI. Requires authentication.",
                "operationId": "4ebc92d554194bc6d144490611ff6698",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "note_id"
                                ],
                                "properties": {
                                    "note_id": {
                                        "description": "The ID of the note to generate flashcards for",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Flashcards created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Flashcards created successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "question": {
                                                        "type": "string",
                                                        "example": "What is photosynthesis?"
                                                    },
                                                    "answer": {
                                                        "type": "string",
                                                        "example": "It is the process by which plants make food using sunlight."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or AI response error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Note not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/flashcards/by-note": {
            "post": {
                "tags": [
                    "Flashcards"
                ],
                "summary": "Get all flashcards for a note",
                "description": "Fetches all flashcards related to a specific note. Requires authentication.",
                "operationId": "83edf4507d5352197052f84a7ddb903f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "note_id"
                                ],
                                "properties": {
                                    "note_id": {
                                        "description": "ID of the note to fetch flashcards for",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Flashcards fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Flashcards fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "note_id": {
                                                        "type": "integer",
                                                        "example": 23
                                                    },
                                                    "question": {
                                                        "type": "string",
                                                        "example": "What is photosynthesis?"
                                                    },
                                                    "answer": {
                                                        "type": "string",
                                                        "example": "It is the process by which plants make food using sunlight."
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}